requirements.txt


django<2
pillow

iamport-rest-client

